+++ src/main/java/ait/cohort46/student/service/StudentServiceImpl.java
@@ -73,8 +73,8 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Integer getStudentsQuantityByNames(Set<String> names) {
-        return (int) StreamSupport.stream(studentRepository.findAll().spliterator(), false)
+    public Long getStudentsQuantityByNames(Set<String> names) {
+        return StreamSupport.stream(studentRepository.findAll().spliterator(), false)
                 .filter(s -> names.stream().anyMatch(n -> n.equalsIgnoreCase(s.getName())))
                 .count();
     }
